Fit-Tracker/pyproject.toml
[project]
name = "fittrack"
version = "0.1.0"
description = "Gamified fitness platform with sweepstakes rewards"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi[all]>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.2.0",

    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "oracledb>=2.0.0",

    # Data Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",

    # Security
    "bcrypt>=4.1.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "cryptography>=42.0.0",

    # Async Tasks
    "celery>=5.3.6",
    "redis>=5.0.1",

    # HTTP Client
    "httpx>=0.26.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.12.0",
    "faker>=22.6.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.98.0",

    # Code Quality
    "ruff>=0.1.15",
    "mypy>=1.8.0",
    "bandit>=1.7.6",
    "pre-commit>=3.6.0",

    # Monitoring
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
    "sentry-sdk>=1.40.0",

    # Load Testing
    "locust>=2.20.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (database required)",
    "e2e: End-to-end tests (full stack required)",
    "slow: Slow running tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "oracledb.*",
    "celery.*",
    "faker.*",
    "factory.*",
]
ignore_missing_imports = true
